teste_unitario/src/main/java/com/example/demo/logic/Calculadora.java
package com.example.demo.logic;

public class Calculadora {
    // Método que soma dois números inteiros e retorna o resultado
    public int somar(int a, int b) {
        return a + b;
    }

    public int subtrair(int a, int b) {
        return a - b;
    }

    public int multiplicar(int a, int b) {
        return a * b;
    }

    public int dividir(int a, int b) {
        if (b == 0) {
            throw new ArithmeticException("Não é possível dividir por zero.");
        }
        return a / b;
    }
}

